multiple classes from FinancialUtils
